Adapt DAP-04 implementation into DAP-02 implementation.
This is needed for upcoming integration tests. Apologies for the
monolithiticity of this PR, there was little time or method to break
this down into more digestible chunks.

With this PR, the following specifications are implemented:
* DAP-02
* VDAF-03 (for "free" with libprio-rs)
* DAP interop test design-02

A few notes:
* taskprov is not yet included. I will backport it next, stacked onto
  this PR.
* The Daphne integration test is disabled; I get an
  `invalidProtocolVersion` error, even when specifically setting
  protocol paths correctly & setting the DAP_DEFAULT_VERSION environment
  variable. I need to investigate more here & possibly reach out to the
  Cloudflare folks if I'm not making a simple error.
* The version of libprio-rs in use is based on a branch; a few minor
  updates are required to adapt the 0.10 branch of libprio-rs to modern
  Janus. I will clean this branch up into a libprio-rs PR once this PR
  has settled (or on request).

+++ integration_tests/src/divviup_api_client.rs
@@ -24,7 +24,7 @@ pub enum ApiVdaf {
         buckets: Vec<u64>,
     },
     Sum {
-        bits: usize,
+        bits: u32,
     },
 }
